Change RightMenu to CustomMenu

+++ Window/BiosView/UefiDecode.cpp
@@ -139,7 +139,7 @@ void BiosViewerWindow::setBiosFvData() {
             return;
         }
         auto fvHeader = (EFI_FIRMWARE_VOLUME_HEADER*)(WindowData->InputImage + offset);
-        INT64 FvLength = fvHeader->FvLength;
+        INT64 FvLength = (INT64)fvHeader->FvLength;
 
         INT64 searchInterval = 0x100;
         INT64 EmptyVolumeLength = 0;
